requirements.txt
# ==================== Core Dependencies ====================
selenium==4.16.0
webdriver-manager==4.0.1

# ==================== Configuration ====================
python-dotenv==1.0.0

# ==================== CLI & UX ====================
rich==13.7.0
click==8.1.7

# ==================== Data Validation ====================
pydantic==2.5.3

# ==================== Utilities ====================
python-dateutil==2.8.2

# ==================== Export (Opcional - requiere compilador C) ====================
# openpyxl==3.1.2
# pandas==2.1.4
# numpy==1.26.4

# ==================== Development (Optional) ====================
# pytest==7.4.3
# black==23.12.1
# flake8==7.0.0
# mypy==1.8.0
